Extract Activity enum
